worker/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pynvml>=11.5.0
psutil>=5.9.0
pyyaml>=6.0
